move WorkspaceChild outside of the model.api package
as it is a business model and not a frameowrk one.
