+++ pantheon-bundle/src/main/java/com/redhat/pantheon/model/document/Document.java
@@ -3,7 +3,7 @@ package com.redhat.pantheon.model.document;
 import com.redhat.pantheon.model.api.Child;
 import com.redhat.pantheon.model.api.Field;
 import com.redhat.pantheon.model.api.FileResource;
-import com.redhat.pantheon.model.api.WorkspaceChild;
+import com.redhat.pantheon.model.workspace.WorkspaceChild;
 
 import javax.annotation.Nonnull;
 import javax.inject.Named;
