+++ pantheon-bundle/src/main/java/com/redhat/pantheon/model/document/DocumentLocale.java
@@ -1,7 +1,7 @@
 package com.redhat.pantheon.model.document;
 
 import com.redhat.pantheon.model.api.Child;
-import com.redhat.pantheon.model.api.WorkspaceChild;
+import com.redhat.pantheon.model.workspace.WorkspaceChild;
 
 public interface DocumentLocale extends WorkspaceChild {
     Child<SourceContent> source();
